backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.32.1
langchain==1.2.13
langgraph==1.1.3
langchain-core==1.2.22
pydantic==2.10.3
python-dotenv==1.2.2
langchain-google-genai==4.2.1
langchain-groq==1.1.2